reformatted files to include all classes in one file
