+++ lib/hotel.rb
@@ -1,4 +1,14 @@
+require "pry"
+
 module Hotel
+  class Reservation
+    def self.all
+      10.times.map do
+        Hotel::Reservation.new
+      end
+    end
+  end
+
   class Room
     def self.all
       return @all_rooms if @all_rooms
@@ -18,19 +28,24 @@ module Hotel
     def self.reservations_on(date)
       Hotel::Room.all.map do |room|
         room if room.reservation_date_range.include?(date)
-      end
-        .compact
+      end.compact
     end
 
+    # def self.all_available_rooms_on(in_date, out_date)
+    #   (Hotel::Room.all - Hotel::Room.reserved_on?(in_date, out_date))
+    # end
+
     attr_reader :reservation_date_range, :room_number
+
+    COST = 200.00
 
     def initialize(room_number)
       @room_number = room_number
     end
 
     def reserve_date_range(in_date, out_date)
-      raise ArgumentError, "Check out date is before check in date" if out_date <= in_date
       @reservation_date_range = in_date...out_date
+      raise ArgumentError, "Check out date is before check in date" if in_date > out_date
     end
 
     def reserved_on?(in_date, out_date)
@@ -38,8 +53,12 @@ module Hotel
       overlap.any?
     end
 
+    def all_available_rooms_on(in_date, out_date)
+      (Hotel::Room.all - Hotel::Room.reservations_on(@reservation_date_range))
+    end
+
     def total_cost
-      (200.00 * @reservation_date_range.to_a.length).round(2)
+      (COST * @reservation_date_range.to_a.length).round(2)
     end
   end
 end
