+++ spec/spec_helper.rb
@@ -8,6 +8,6 @@ require "minitest/reporters"
 
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
-require_relative "../lib/room"
-require_relative "../lib/reservation"
-# require_relative "../lib/hotel"
+# require_relative "../lib/room"
+# require_relative "../lib/reservation"
+require_relative "../lib/hotel"
